refactor: rename src/commands to src/phantom/command
- Move all command files from src/commands/ to src/phantom/command/
- Update all import paths to reflect the new directory structure
- Update relative imports in moved files to account for additional directory level
- All tests pass and linting is clean

Closes #31
