+++ src/phantom/command/exec.ts
@@ -1,14 +1,21 @@
 import { spawn } from "node:child_process";
 import { exit } from "node:process";
-import { whereGarden } from "../gardens/commands/where.ts";
+import { whereGarden } from "../../gardens/commands/where.ts";
 
-export async function shellInGarden(gardenName: string): Promise<{
+export async function execInGarden(
+  gardenName: string,
+  command: string[],
+): Promise<{
   success: boolean;
   message?: string;
   exitCode?: number;
 }> {
   if (!gardenName) {
     return { success: false, message: "Error: garden name required" };
+  }
+
+  if (!command || command.length === 0) {
+    return { success: false, message: "Error: command required" };
   }
 
   // Validate garden exists and get its path
@@ -18,25 +25,18 @@ export async function shellInGarden(gardenName: string): Promise<{
   }
 
   const gardenPath = gardenResult.path as string;
-  // Use user's preferred shell or fallback to /bin/sh
-  const shell = process.env.SHELL || "/bin/sh";
+  const [cmd, ...args] = command;
 
   return new Promise((resolve) => {
-    const childProcess = spawn(shell, [], {
+    const childProcess = spawn(cmd, args, {
       cwd: gardenPath,
       stdio: "inherit",
-      env: {
-        ...process.env,
-        // Add environment variable to indicate we're in a phantom garden
-        PHANTOM_GARDEN: gardenName,
-        PHANTOM_GARDEN_PATH: gardenPath,
-      },
     });
 
     childProcess.on("error", (error) => {
       resolve({
         success: false,
-        message: `Error starting shell: ${error.message}`,
+        message: `Error executing command: ${error.message}`,
       });
     });
 
@@ -44,7 +44,7 @@ export async function shellInGarden(gardenName: string): Promise<{
       if (signal) {
         resolve({
           success: false,
-          message: `Shell terminated by signal: ${signal}`,
+          message: `Command terminated by signal: ${signal}`,
           exitCode: 128 + (signal === "SIGTERM" ? 15 : 1),
         });
       } else {
@@ -58,26 +58,16 @@ export async function shellInGarden(gardenName: string): Promise<{
   });
 }
 
-export async function shellHandler(args: string[]): Promise<void> {
-  if (args.length < 1) {
-    console.error("Usage: phantom shell <garden-name>");
+export async function execHandler(args: string[]): Promise<void> {
+  if (args.length < 2) {
+    console.error("Usage: phantom exec <garden-name> <command> [args...]");
     exit(1);
   }
 
   const gardenName = args[0];
+  const command = args.slice(1);
 
-  // Get garden path for display
-  const gardenResult = await whereGarden(gardenName);
-  if (!gardenResult.success) {
-    console.error(gardenResult.message);
-    exit(1);
-  }
-
-  // Display entering message
-  console.log(`Entering garden '${gardenName}' at ${gardenResult.path}`);
-  console.log("Type 'exit' to return to your original directory\n");
-
-  const result = await shellInGarden(gardenName);
+  const result = await execInGarden(gardenName, command);
 
   if (!result.success) {
     if (result.message) {
@@ -86,6 +76,6 @@ export async function shellHandler(args: string[]): Promise<void> {
     exit(result.exitCode ?? 1);
   }
 
-  // Exit with the same code as the shell
+  // For successful commands, exit with the same code as the child process
   exit(result.exitCode ?? 0);
 }
